refactor : gui, core 패키지 구분
