+++ src/main/java/com/zuniorteam/racingcar/core/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package com.zuniorteam.racingcar.strategy;
+package com.zuniorteam.racingcar.core.strategy;
 
 @FunctionalInterface
 public interface MovingStrategy {
