+++ src/main/java/com/zuniorteam/racingcar/core/strategy/RandomMovingStrategy.java
@@ -1,0 +1,9 @@
+package com.zuniorteam.racingcar.core.strategy;
+
+public class RandomMovingStrategy implements com.zuniorteam.racingcar.core.strategy.MovingStrategy {
+
+    @Override
+    public boolean isMoving() {
+        return false;
+    }
+}
