--- src/main/java/com/zuniorteam/racingcar/strategy/RandomMovingStrategy.java
@@ -1,9 +1,0 @@
-package com.zuniorteam.racingcar.strategy;
-
-public class RandomMovingStrategy implements MovingStrategy {
-
-    @Override
-    public boolean isMoving() {
-        return false;
-    }
-}
