+++ src/test/java/com/zuniorteam/racingcar/core/CarTest.java
@@ -1,7 +1,7 @@
-package com.zuniorteam.racingcar;
+package com.zuniorteam.racingcar.core;
 
-import com.zuniorteam.racingcar.strategy.RandomMovingStrategy;
-import org.junit.jupiter.api.Assertions;
+import com.zuniorteam.racingcar.core.Car;
+import com.zuniorteam.racingcar.core.strategy.RandomMovingStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
